requirements.txt
pandas==1.5.0
streamlit==1.25.0
transformers==4.31.0
llama-index==0.7.2
streamlit-chat
